+++ APIScripts/IP_Address API/readme.md
@@ -1,0 +1,15 @@
+
+##A simple python program that generates information about a particualar IP Address with the help of 'ip-api' API
+
+**Imported modules:**
+-requests
+-simple_colors
+
+**Website URL** : https://ip-api.com/
+
+**API Documentation URL** : https://ip-api.com/docs
+
+------------
+
+##Author:
+Debashish Kundu (https://www.github.com/its-me-debk007)
