+++ APIScripts/IP_Address API/requirements.txt
@@ -1,0 +1,6 @@
+
+Need the following modules:
+
+-requests
+
+-simple_colors
